Added detox smoke test's in the travis CI
