+++ .travis.yml
@@ -1,7 +1,44 @@
-language: node_js
-dist: precise
-node_js:
-  - 6.0
+language: objective-c
+osx_image: xcode9
+xcode_sdk: iphonesimulator11.0
+sudo: required
+env:
+  global:
+    - NODE_VERSION=stable
+    - COVERALLS_REPO_TOKEN=4eYQDtWoBJlDz2QkxoQ2UcnmJFcOB7zkv
+
+install:
+  - brew update
+  - sudo launchctl limit maxfiles 2048 unlimited
+  - sudo ulimit -n 10000
+  - curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.33.1/install.sh | bash
+  - export NVM_DIR="$HOME/.nvm" && [ -s "$NVM_DIR/nvm.sh" ] && . "$NVM_DIR/nvm.sh"
+  - nvm install $NODE_VERSION
+  - nvm use $NODE_VERSION
+  - brew install watchman
+  - brew install yarn
+  - brew tap facebook/fb
+  # START HACK This is a hack since the homebrew version updates daily. we install a
+  # specific working version. In the future install a version of fbsimctl
+  # that works.
+  - git clone https://github.com/facebook/FBSimulatorControl.git
+  - cd FBSimulatorControl && git checkout 2a461c4e31f12b29974a5a6d184126c2526c62d1 && ./build.sh fbsimctl build output && cd ../
+  - mv FBSimulatorControl/output/Frameworks/* /usr/local/Frameworks/
+  - mv FBSimulatorControl/output/bin/* /usr/local/bin/
+  # END HACK
+  #  Add those lines again once the hack is removed. Remove the --HEAD flag.
+  #- brew tap facebook/fb
+  #- export CODE_SIGNING_REQUIRED=NO && brew install fbsimctl --HEAD
+  - export CODE_SIGNING_REQUIRED=NO
+  - brew tap wix/brew
+  - brew install applesimutils --HEAD
+  - yarn install
+
+
+before_script:
+  # - travis_wait yarn run build:release:ios
+  - fbsimctl shutdown
+
 notifications:
   email: false
   webhooks:
@@ -9,9 +46,20 @@ notifications:
       - https://zulip.org/zulipbot/travis
     on_success: always
     on_failure: always
-before_install:
-  - npm install
-install:
-  npm run test
-env:
-  - COVERALLS_REPO_TOKEN=4eYQDtWoBJlDz2QkxoQ2UcnmJFcOB7zkv
+
+script:
+  - yarn run build:release-ios
+  - yarn run test
+  - yarn run ui-test
+
+cache:
+  yarn: true
+  directories:
+    - node_modules
+    - $HOME/Library/Caches/Homebrew
+
+addons:
+  postgresql: "9.3"
+  apt:
+    packages:
+      - iproute2
