+++ package.json
@@ -21,11 +21,12 @@
     "test:lint": "eslint src/",
     "test:flow": "flow",
     "test:flow-coverage": "flow-coverage-report -i 'src/**/*.js' -x '**/__tests__/**' -t html",
-    "test:unit": "jest",
-    "test:e2e": "detox build; jest -t '__e2e__tests__/basic.js'",
+    "test:unit": "jest src/",
+    "test:e2e": "jest -t '__e2e__tests__/basic.spec.js'",
     "test:coveralls": "jest --coverage && cat ./coverage/lcov.info | coveralls",
     "test:prettier": "prettier-check  --single-quote --jsx-bracket-same-line --trailing-comma all --print-width 100 --parser flow `find src -name \"*.js\"`",
-    "test": "npm run test:lint && npm run test:prettier && npm run test:coveralls && npm run test:flow"
+    "test": "npm run test:lint && npm run test:prettier && npm run test:coveralls && npm run test:flow",
+    "build:release-ios": "detox build --configuration ios.sim.release > /dev/null"
   },
   "dependencies": {
     "@expo/react-native-action-sheet": "1.0.0",
@@ -106,13 +107,14 @@
     "setupTestFrameworkScriptFile": "./jestSetup.js",
     "transformIgnorePatterns": [
       "node_modules/(?!react-native|@expo/react-native-action-sheet|react-navigation)"
-    ]
+    ],
+    "testPathIgnorePatterns": ["/node_modules/", "/__e2e__tests__/"]
   },
   "detox": {
     "configurations": {
-      "ios.sim.debug": {
-        "binaryPath": "ios/build/Build/Products/Debug-iphonesimulator/ZulipMobile.app",
-        "build": "xcodebuild -project ios/ZulipMobile.xcodeproj -scheme ZulipMobile -configuration Debug -sdk iphonesimulator -derivedDataPath ios/build",
+      "ios.sim.release": {
+        "binaryPath": "ios/build/Build/Products/Release-iphonesimulator/ZulipMobile.app",
+        "build": "xcodebuild -project ios/ZulipMobile.xcodeproj -scheme ZulipMobile -configuration Release -sdk iphonesimulator -derivedDataPath ios/build",
         "type": "ios.simulator",
         "name": "iPhone 6"
       }
